Correct the misplaced recv in ipmi_lan_recv_packet() in lan and lanplus interfaces.
Patch is submitted by Ferenc Wagner (wferi at niif dot hu). Quoting his
email:

"The rationale is that the following code expects to use the return
value of recv (ie. a byte count) not that of select (the number of file
descriptors). recv should be executed anyway, that's the whole point of
this snippet..."

The original patch is for lan interface but it also applies to lanplus.