ID:446 - Fix broken firewall reset iterator
The netfn structures are populated sequentially in the _gather_info
function and they are accessed only by even indices (and to the double
of its real length). This results in a segmentation fault if you run

* ipmitool firewall reset

This patch fixes the behaviour so that the ipmi_firewall_reset function
treats the structure the same way as the rest of the code does.

Signed-off-by: Boris Ranto <branto@redhat.com>