raw: allow for responses larger than 256 bytes
All the interfaces use a static struct ipmi_rs to capture a response.
The struct has an embedded buffer that is a fixed size for all
interfaces, regardless of the interface's max transfer size. Bumping up
the buffer size to the maximum of all transfer sizes would increase
memory usage by too much, so this change sets a static buffer for each
of the interfaces that is the right size for that interface.

This does increase the BSS slightly, but only by ~57kB (in my build).

Resolves ipmitool/ipmitool#1

Signed-off-by: Vernon Mauery <vernon.mauery@intel.com>