ignore sol keepalive timeout when keepalives fail
When a keepalive is requested prior to keepalive timeout,
the keepalive request is flagged as successful without sending
the request over wire. If BMC goes dark, after a SOL session
has been established, this fake success from keepalive request
can cause the ipmi_sol_red_pill to get stuck in a forever loop.
The fix is to not skip the keepalive requests after one of them
fails until there is a pass.