lib/ipmi_ekanalyzer: cleanup variable reassignment
In a few methods there are assignments to a return value that is
ignored.  Convert the code to cast to void to indicate that the return
value is deliberately ignored.

[lib/ipmi_ekanalyzer.c:2556, 2561, 2566, 2571,
                       2665, 2910, 2914, 2918,
		       2922, 2926, 2930]: (style)
Variable 'ret' is reassigned a value before the old
one has been used.

[lib/ipmi_ekanalyzer.c:2934, 2669]: (style) Variable
'ret' is assigned a value that is never used.

[lib/ipmi_ekanalyzer.c:2671, 2936]: (style) Variable
'file_offset' is assigned a value that is never used.

[lib/ipmi_ekanalyzer.c:3315]: (style) Variable
'ioffset' is assigned a value that is never used.

Signed-off-by: Patrick Venture <venture@google.com>