Fix options= parsing for persistent and efiboot
When using 'persistent' and/or 'efiboot' in an 'options=' like this:
  ipmitool -I lanplus chassis bootdev disk options=efiboot,persistent

Allow them to be used together. Previously if you specified them
together only the last one specified would take effect. That would
mean you could either specify to set for EFI or you could set it to be
persistent. But could not set EFI and persistent.

Changed the mask values so that they no longer wipe out previous
values.

Resolves: ipmitool#163