read_fru_area: fix memset() parameters
Current memset parameters are calculated in the assumption
that we use the buffer for whole FRU. Whereas we use individual
buffers for each area that are smaller than whole FRU size.
That may lead to attempt to memset beyond allocated buffer.
Also: add comment to fail path in fru_area_print_product().