Clamp, instead of roll over, analog sensor byte
When convering a double value to an integer byte, in the IPMI SDR,
there was previously no bounds-checking, causing the byte to
roll over when an out-of-range value was passed in.

Instead of rolling over, clamp the byte to the appropriate minimum
or maximum, based on the type (signed versus unsigned) of the
byte. This behavior, to me, seems safer and more correct.

Signed-off-by: Josh Lehan <krellan@google.com>