Support multiple lanplus
We require holding multiple lanplus interfaces open at once, our approach is to make copies of the lanplus_intf, and supply them as needed. To do so we must remove the global state.

We'd like to add void* pointers where appropriate to the ipmi_intf struct and move global state into heap-allocated structs that are attached to individual ipmi_intf instances. Doing so makes the library a tad bit more thread-safe as a side-effect.